style fixed

+++ src/main/resources/init_db.sql
@@ -2,5 +2,5 @@
 CREATE TABLE `books` (
     `id` INT NOT NULL AUTO_INCREMENT,
     `title` VARCHAR(255),
-    `price` INT,
+    `price` DECIMAL,
     PRIMARY KEY (`id`));
